Passando criação do DB para a fase
