+++ SQLGame/Assets/Scripts/Levels/fase1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b0799b7baded486c9af8a3fefaa3319
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
